Allow the assert to be redefined
The C code is very clean but it depends on the system assert which can print and abort the process. This is not always acceptable, so this change defines a `ASSERT` macro that can be redefined to something else to avoid the dependency on the system `assert`, even while debugging.